Expression __hash__ implementation.
It works by recursively taking the hash() of every children and mixing that with sort information, operation kind, variable names and value if present.

If the Expression is commutative, it sort() the children to retrieve a kind of canonicalized hash of the expression.
This way, a dictionary with two keys that are different versions of a same expression but written in different ways would point to the same value.
This is a very handy property for memory maps.

Example:
    anded=(((bv1 & bv2) | 0x12345678) + 0xbababebe)
    anded2=(0xbababebe + (0x12345678 | (bv2 & bv1)))

    hash(anded) == hash(anded2)