changed __backend__ to be a function.
This way we can optimize the function away once we calculate the solver expression and replace it with a lambda that returns the computed value.
Also changed the DeferredMemoryRead Z3 implementation to really defer the creation of the Array entries until the __backend__() method of the
instance is called. This way, even if a DeferredMemoryRead instance is created, the array is only created / updated when the solver needs access to
it. Giving space to further optimizations on the pre-processing to take place (avoiding the creation of the array at all for example).

Example:

x = ReadDword(reg1)
x &= 0xffff

Before we had to create the 4 memory reads, just to drop them one sentence after, with the new way, only the BvExprs are created and the solver is not touched at all.